move function generate_random_values to a .cpp
