+++ src/gather/single_threaded/benchmark_agg_avx_64bit.cpp
@@ -20,28 +20,7 @@ using namespace std;
 #include "measures.h"
 struct measures scalar, linear, gather, seti;
 
-
-template <typename T>
-void generate_random_values(T* array, uint64_t number) {
-  static_assert(is_integral<T>::value, "Data type is not integral.");
-  std::random_device rd;
-  std::mt19937::result_type seed = rd() ^ (
-          (std::mt19937::result_type)
-          std::chrono::duration_cast<std::chrono::seconds>(
-          std::chrono::system_clock::now().time_since_epoch()
-          ).count() +
-          (std::mt19937::result_type)
-          std::chrono::duration_cast<std::chrono::microseconds>(
-          std::chrono::high_resolution_clock::now().time_since_epoch()
-          ).count());
-
-  std::mt19937 gen(seed);
-  std::uniform_int_distribution<T> distrib(1, 6);
-
-  for (uint64_t j = 0; j < number; ++j) {
-    array[j] = distrib(gen);
-  }
-}
+#include "generate_random_values.cpp"
 
 bool benchmark(measures* res, uint64_t correct_result, const uint64_t* values, uint64_t n, const uint32_t stride, double GB, function<uint64_t(const uint64_t*,uint64_t, const uint32_t)> func) {
